Add -M4294967295 to zstd options.
Without that option, zstd will refuse to decompress archives which have
been created with option --long=28 or larger (for testing, one must compress
from a pipe, because zstd shrinks the --long value according to the filesize
if no pipe is used).

The huge number passed covers the maximal possible value --long=31.
Usage of the huge number is safe if the file can be decompressed with less
memory: The memory allocated by zstd is only large if that amount is really
*needed* to decompress the file. For instance, if you decompress a _short_
file generated with cat ... | zstd --ultra -22 --long=31 >...
also only a small amount of memory is allocated, independent of the -M option.