Make wvText use /dev/stdout instead of -.
wvText is a script that uses > "${2}" shell redirection to point to its
output file. Its shebang is /bin/sh, which is often a symlink to bash,
which does not reinterpret > - as going to stdout.

But random people on the internet seem to be saying that while
/dev/stdout is not universal (not POSIX, and absent on e.g. MinGW and
AIX), bash knows to emulate it when it's passed as a command line
parameter:

https://unix.stackexchange.com/questions/36403/

So I'm blindly guessing that this improves the chances of correctly
handling a Word doc where abiword is installed (which pulls in wvText,
which is the first Word preprocessor attempted).