Consider also `lessfilter` in user PATH
In order to not pollute their homedir, many users don't like to drop
all sort of configs/binaries like hidden `.lessfilter` into their home.

One way to support a user custom `lessfilter`, is looking if that's
available in their PATH, e.g. `~/bin/` or `~/.local/bin/`. This would
also give the possibility to get the `lessfilter` as a shell plugin,
without messing user's homedir.

This adds support for that, but still giving priority to `~/.lessfilter`
when found.