add configure option to always install all completions
This is useful for distro packaging. Many distros have a policy that
shell completions shall always be installed, even if the shell they are
for is not installed at the time.

This is useful because buildbot chroots often don't have additional
shells installed, so users will *never* get zsh completions.