Fix usage instructions for bourne shell (sh/bash).
The fix is self-evident as a matter of syntax. With the previous
instruction to `eval lesspipe.sh`, the shell will execute `lesspipe.sh`
but simply echo its result to stdout. To set `LESSOPEN` the output
needs to be eval'd, so capture it with the shell's $() command
substitution.

Demonstration:

```
$ unset LESSOPEN
$ eval lesspipe.sh
LESSOPEN="|/usr/bin/lesspipe.sh %s"
export LESSOPEN
$ echo $LESSOPEN

$ eval "$(lesspipe.sh)"
$ echo $LESSOPEN
|/usr/bin/lesspipe.sh %s
```

Tested with GNU bash, version 5.1.16 (and a busybox `sh`).